+++ src/main/java/org/smileyface/commands/music/LeaveCommand.java
@@ -2,9 +2,12 @@ package org.smileyface.commands.music;
 
 import java.util.Objects;
 import java.util.Set;
+import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.channel.middleman.AudioChannel;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
+import org.smileyface.audio.MusicManager;
+import org.smileyface.audio.TrackQueue;
 import org.smileyface.checks.Checks;
 import org.smileyface.checks.CommandFailedException;
 import org.smileyface.commands.BotCommand;
@@ -19,11 +22,14 @@ public class LeaveCommand extends BotCommand {
 
     @Override
     public void run(SlashCommandInteractionEvent event) throws CommandFailedException {
-        AudioChannel audioChannel = Checks.authorInVoice(
-                Objects.requireNonNull(event.getMember()));
+        Member author = Objects.requireNonNull(event.getMember());
+        AudioChannel audioChannel = Checks.authorInVoice(author);
         Checks.botConnectedToAuthorVoice(audioChannel);
 
-        Music.leaveVoiceIfConnected(Objects.requireNonNull(event.getGuild()));
+        TrackQueue queue = MusicManager.getInstance().getQueue(author.getGuild().getIdLong());
+        Music.leaveVoiceIfConnected(author.getGuild());
+        queue.getTrackQueueEmbed().setLastCommand(author, "Stopped the music, "
+                        + "and made the bot leave voice channel");
         event.reply("Left channel!").setEphemeral(true).queue();
     }
 }